requirements.txt
PyQt6>=6.6
bleak>=0.20.0
pyserial>=3.5
h5py>=3.8
numpy>=1.24
matplotlib>=3.7
sounddevice>=0.4.6
pyttsx3>=2.90
sqlalchemy>=2.0
